learn

+++ hw1/synthesizer/GuitarString.java
@@ -13,10 +13,10 @@ public class GuitarString {
 
     /* Create a guitar string of the given frequency.  */
     public GuitarString(double frequency) {
-        int capacity = (int)Math.round(SR/frequency);
+        int capacity =(int) Math.round(SR / frequency);
         buffer = new ArrayRingBuffer<>(capacity);
         while (!buffer.isFull()) {
-            buffer.enqueue((double)0);
+            buffer.enqueue(0.0);
         }
     }
     /* Pluck the guitar string by replacing the buffer with white noise. */
